Finish restructuring of layout to use component structure.
