+++ src/components/LayoutManager/components/Container/Container.scss
@@ -1,24 +1,24 @@
-.hoverDiv:hover {
-    opacity:50%;
-}
+
 
 .handleBarHorizontal {
   height:100%;
   width: 3px;
+    background-color: #414141;
 }
 
 .handleBarHorizontal:hover {
-    background-color: blue;
+    background-color: #007acc;
     cursor: ew-resize;
 }
 
 .handleBarVertical {
   width:100%;
   height: 3px;
+    background-color: #414141;
 }
 
 .handleBarVertical:hover {
-    background-color: blue;
+    background-color: #007acc;
     cursor: ns-resize;
 }
