+++ src/components/LayoutManager/components/PlaceHolder/PlaceHolder.scss
@@ -1,0 +1,3 @@
+.hoverDiv:hover {
+    opacity:70%;
+}
